SDL: Adding Input Handling and CollisionSystem
This change adds primitive input handling and implementation
of AABB Collision System.

+++ sources/game/src/main.cpp
@@ -3,8 +3,8 @@
 #include <stdint.h>
 
 int main(int argc, char** argv) {
-    const int FPS = 60;
-    const int frameDelay = 1000 / FPS;
+    const int FPS = 60 ;
+    const float frameDelay = 1000 / FPS;
 
     uint32_t frameStart;
     int frameTime;
@@ -17,8 +17,7 @@ int main(int argc, char** argv) {
 		game->handleEvents();
 		game->update();
 		game->render();
-
-		frameTime = SDL_GetTicks() - frameStart;
+        frameTime = SDL_GetTicks() - frameStart;
 		if (frameDelay > frameTime) {
             SDL_Delay(frameDelay - frameTime);
 		}
